Move session middleware in to session package
