+++ src/session/destroySession.js
