+++ test/session/createSession.test.js
@@ -1,7 +1,7 @@
 const { expect, sinon } = require('../util/chai');
-const createSession = require('../../src/middleware/createSession');
+const createSession = require('../../src/session/createSession');
 
-describe('middleware/createSession', () => {
+describe('session/createSession', () => {
   it('calls req.session.generate to create a session', done => {
     const req = { session: { generate: sinon.stub() } };
     const assertions = () => {
